Move ErrorDialog to commons
