+++ src/js/components/common/error-dialog.scss
@@ -3,7 +3,6 @@
 
 .dialog {
   width: 400px;
-  min-height: 300px;
   background-color: white;
   margin: 2rem auto;
   box-shadow: 0px 3px 5px $minard-shadow;
@@ -36,6 +35,7 @@
   @include nitti-grotesk-light;
   @include font-size-M;
   padding: 1.5rem;
+  min-height: 168px;
 
   :first-child {
     margin-top: 0;
