--- src/js/components/signup-view/index.scss
@@ -1,36 +1,0 @@
-@import '../mixins';
-@import '../definitions';
-
-.header {
-  height: 64px;
-  margin: 0;
-
-  @include nitti-grotesk-medium;
-  @include font-size-M;
-  color: white;
-}
-
-.header-background {
-  background-image: linear-gradient(to left, #17b7b7, #54c6a8);
-}
-
-.logo {
-  text-align: center;
-}
-
-.minard {
-  @include nitti-grotesk-bold;
-  @include font-size-XXL;
-
-  display: inline-block;
-  margin: 0;
-  transition: transform 0.6s ease-in-out;
-  transform-origin: 50% 55%;
-  transform: rotate(0deg);
-  user-select: none;
-  cursor: default;
-
-  &:hover {
-    transform: rotate(360deg);
-  }
-}
